requirements.txt
anthropic>=0.25.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.126.0
PyGithub>=2.3.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
requests>=2.31.0
resend>=2.0.0
python-dotenv>=1.0.1
pyyaml>=6.0.1
python-frontmatter>=1.1.0
tenacity>=8.2.3
sqlite-utils>=3.36
